refactor: reconciled changes from context branch into thunks
