+++ shared/components/AuthButton/AuthButton.tsx
@@ -1,35 +1,38 @@
 import React from 'react';
 
-type AuthButtonProps = {
-  isLoggedIn: boolean;
-  userName: string;
-  login(): void;
-  logout(): void;
-};
+import { useAuth, useUserName } from '../../../store/auth';
+import { useRenderCount } from '../../../utils/useRenderCount';
 
-export const AuthButton = ({ isLoggedIn, userName, login, logout }: AuthButtonProps) => (
-  <>
-    <div>
-      <div className={'button-container'}>
-        <button type={'button'} onClick={() => (isLoggedIn ? logout() : login())}>
-          {isLoggedIn ? 'Logout' : 'Login'}
-        </button>
+export const AuthButton = () => {
+  const { loggedIn, login, logout } = useAuth();
+  const userName = useUserName();
+  const renderCount = useRenderCount();
+
+  return (
+    <>
+      <div>
+        <div className={'button-container'}>
+          <button type={'button'} onClick={() => (loggedIn ? logout() : login())}>
+            {loggedIn === 'unknown' ? 'Login' : loggedIn ? 'Logout' : 'Login'}
+          </button>
+        </div>
+        {loggedIn !== 'unknown' && <div className={'user'}>User: {userName ?? '<empty>'}</div>}
+        (simulated 250ms latency on all requests)
+        {renderCount && `AuthButton render count: ${renderCount}`}
       </div>
-      <div className={'user'}>User: {userName ?? '<empty>'}</div>
-      (simulated 250ms latency on all requests)
-    </div>
-    <style jsx>{`
-      div {
-        display: flex;
-        width: 740px;
-      }
-      .button-container {
-        width: 150px;
-      }
+      <style jsx>{`
+        div {
+          display: flex;
+          width: 740px;
+        }
+        .button-container {
+          width: 150px;
+        }
 
-      .user {
-        flex: 1;
-      }
-    `}</style>
-  </>
-);
+        .user {
+          flex: 1;
+        }
+      `}</style>
+    </>
+  );
+};
